feat: continuing /begin command

+++ gradykrueger/bot/bot.py
@@ -1,8 +1,10 @@
 import os
+from pathlib import Path
 
 import cv2 as cv
 from django.conf import settings
-from telebot import TeleBot, types
+from telebot import TeleBot
+from telebot.types import BotCommand, MenuButtonCommands, Message
 
 from .utils import *
 
@@ -10,12 +12,12 @@ bot = TeleBot(settings.BOT_TOKEN)
 imgs_path = os.path.join(settings.BASE_DIR.__str__() + "/bot/imgs/")
 
 # menu commands
-start_command = types.BotCommand(command="start", description="Старт бота")
-help_command = types.BotCommand(command="help", description="Помощь")
-begin_checking_command = types.BotCommand(
+start_command = BotCommand(command="start", description="Старт бота")
+help_command = BotCommand(command="help", description="Помощь")
+begin_checking_command = BotCommand(
     command="begin", description="Начать проверку работ"
 )
-create_template_command = types.BotCommand(
+create_template_command = BotCommand(
     command="create_template", description="Сделать шаблон теста"
 )
 
@@ -55,8 +57,8 @@ test_config = Config()
 
 # message handlers
 @bot.message_handler(commands=["start", "help"])
-def send_welcome(message):
-    bot.set_chat_menu_button(message.chat.id, types.MenuButtonCommands("commands"))
+def send_welcome(message: Message):
+    bot.set_chat_menu_button(message.chat.id, MenuButtonCommands("commands"))
     bot.send_message(
         message.chat.id,
         "Старт или помощь",
@@ -64,7 +66,7 @@ def send_welcome(message):
 
 
 @bot.message_handler(commands=["begin"])
-def start_checking(message):
+def start_checking(message: Message):
     bot.send_message(
         message.chat.id,
         "Вопросы в тесте могут содержать несколько правильных ответов?\nОтвет засчитывается за 0,5 баллов, если пропущен только один ответ.(Да/Нет)",
@@ -74,15 +76,15 @@ def start_checking(message):
 
 
 @bot.message_handler(commands=["create_template"])
-def create_template(message):
+def create_template(message: Message):
     bot.send_message(message.chat.id, "Выберите количество колон на листе.")
 
     bot.register_next_step_handler(message, choose_columns)
 
 
-def choose_is_multiple_answer(message):
-    if message.text.lower() in ["да", "нет"]:
-        test_config.set_multiple_answer(message.text.lower() == "да")
+def choose_is_multiple_answer(message: Message):
+    if str(message.text).lower() in ["да", "нет"]:
+        test_config.set_multiple_answer(str(message.text).lower() == "да")
 
         msg = bot.send_message(message.chat.id, "Выберите количество колон на листе.")
 
@@ -96,8 +98,8 @@ def choose_is_multiple_answer(message):
         bot.register_next_step_handler(msg, choose_is_multiple_answer)
 
 
-def choose_columns(message):
-    if message.text.isnumeric() and int(message.text) > 0:
+def choose_columns(message: Message):
+    if str(message.text).isnumeric() and int(str(message.text)) > 0:
         test_config.set_columns(int(str(message.text)))
 
         msg = bot.send_message(message.chat.id, "Выберите количество строк на листе.")
@@ -115,8 +117,8 @@ def choose_columns(message):
         bot.register_next_step_handler(msg, choose_columns)
 
 
-def choose_rows(message):
-    if message.text.isnumeric() and int(message.text) > 0:
+def choose_rows(message: Message):
+    if str(message.text).isnumeric() and int(str(message.text)) > 0:
         test_config.set_rows(int(str(message.text)))
 
         msg = bot.send_message(
@@ -139,7 +141,7 @@ def choose_rows(message):
 def choose_answers_quantity(
     message,
 ):
-    if message.text.isnumeric() and int(message.text) > 0:
+    if str(message.text).isnumeric() and int(str(message.text)) > 0:
         test_config.set_n(int(str(message.text)))
 
         # if is_multiple_answer is None, then user is creating test template and not checking tests
@@ -155,14 +157,16 @@ def choose_answers_quantity(
                     open(imgs_path + "test_template.jpg", "rb"),
                     "Вот ваш шаблон для тестов!",
                 )
-
             except:
                 bot.send_message(
                     message.chat.id,
                     "Во время генерации шаблона произошла ошибка! Попробуйте создать его снова.",
                 )
                 print("Something went wrong during creating template!")
-            os.remove(imgs_path + "test_template.jpg")
+
+            if Path(os.path.join(imgs_path + "test_template.jpg")).is_file():
+                os.remove(imgs_path + "test_template.jpg")
+
             test_config.clear()
         else:
             msg = bot.send_message(
@@ -183,12 +187,12 @@ def choose_answers_quantity(
         )
 
 
-def proccess_correct_answers(message: types.Message):
+def proccess_correct_answers(message: Message):
     if message.photo is not None:
         # user send a message
         try:
-            fileID = message.photo[-1].file_id
-            file_info = bot.get_file(fileID)
+            file_id = message.photo[-1].file_id
+            file_info = bot.get_file(file_id)
             downloaded_file = bot.download_file(file_info.file_path)
 
             f = open(imgs_path + "test_correct_received.jpg", "wb")
@@ -214,25 +218,89 @@ def proccess_correct_answers(message: types.Message):
 
             cv.imwrite(imgs_path + "test_correct.jpg", transformed_correct)
 
-            bot.send_photo(
+            msg = bot.send_photo(
                 message.chat.id,
                 open(imgs_path + "test_correct.jpg", "rb"),
-                "Вот правильные ответы на ваш тест!",
+                "Вот правильные ответы на ваш тест! Правильные ли ответы, выделенные зеленым ? (Да/Нет)",
             )
 
+            bot.register_next_step_handler(msg, confirm_test_answers)
         except:
             bot.send_message(
                 message.chat.id,
                 "Во время обработки изображения произошла ошибка! Попробуйте снова.",
             )
-        test_config.clear()
-        os.remove(imgs_path + "test_correct.jpg")
-        os.remove(imgs_path + "test_correct_received.jpg")
+            print("Error during image proccesing!")
+
+        if Path(os.path.join(imgs_path + "test_correct.jpg")).is_file():
+            os.remove(imgs_path + "test_correct.jpg")
+
+        if Path(os.path.join(imgs_path + "test_correct_received.jpg")).is_file():
+            os.remove(imgs_path + "test_correct_received.jpg")
+
     else:
         bot.send_message(
             message.chat.id,
-            "Не понял вашего ответа, попробуйте еще раз!\nОтправьте фотографию шаблона c заполненными ответами или отправьте только одну фотографию.",
+            "Не понял вашего ответа, попробуйте еще раз!\nОтправьте фотографию шаблона c заполненными ответами, отправьте только одну фотографию.",
         )
+
+        bot.register_next_step_handler(
+            message,
+            proccess_correct_answers,
+        )
+
+
+def confirm_test_answers(message: Message):
+    if str(message.text).lower() == "да":
+        msg = bot.send_message(
+            message.chat.id,
+            "Пришлите сюда фотографии заполненных тестов, которые нужно проверить !",
+        )
+
+        bot.register_next_step_handler(msg, check_tests)
+    elif str(message.text).lower() == "нет":
+        msg = bot.send_message(
+            message.chat.id,
+            "Отправьте фотографию с правильными ответами! Если у вас нет шаблона, то воспользуйтесь командой /create_template",
+        )
+
+        bot.register_next_step_handler(msg, proccess_correct_answers)
+    else:
+        msg = bot.send_message(
+            message.chat.id,
+            "Не понял вашего вопроса, попробуйте еще раз!\nОтветьте Да или Нет.",
+        )
+
+        bot.register_next_step_handler(
+            msg,
+            confirm_test_answers,
+        )
+
+
+def check_tests(message: Message):
+    if message.photo is not None:
+        # interate over all received photos
+        for photo in message.photo[1:]:
+            try:
+                file_id = photo.file_id
+                file_info = bot.get_file(file_id)
+                downloaded_file = bot.download_file(file_info.file_path)
+
+                f = open(imgs_path + "test_marked_received.jpg", "wb")
+                f.write(downloaded_file)
+
+                marked_answers_image = cv.imread(imgs_path + "test_marked_received.jpg")
+            except:
+                print("Photo could not be proccessed!")
+
+        test_config.clear()
+    else:
+        msg = bot.send_message(
+            message.chat.id,
+            "Не понял вашего ответа, попробуйте еще раз!\nОтправьте фотографии тестов c заполненными ответами, отправьте только одну фотографию.",
+        )
+
+        bot.register_next_step_handler(msg, check_tests)
 
 
 bot.infinity_polling()
